Completed mobile view and extracted components
